refactor: 사용자 포인트 지급 서비스 로직 이동
