+++ src/test/java/org/example/oddventure/domain/admin/unit/AdminUserServiceTest.java
@@ -8,11 +8,8 @@ import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 import org.example.oddventure.common.exception.GlobalException;
-import org.example.oddventure.domain.admin.dto.request.PointAdjustRequest;
-import org.example.oddventure.domain.admin.dto.response.PointAdjustResponse;
 import org.example.oddventure.domain.admin.dto.response.UserAdminResponse;
 import org.example.oddventure.domain.admin.service.AdminUserService;
-import org.example.oddventure.domain.match.repository.MatchRepository;
 import org.example.oddventure.domain.user.entity.User;
 import org.example.oddventure.domain.user.repository.UserRepository;
 import org.junit.jupiter.api.DisplayName;
@@ -31,9 +28,6 @@ public class AdminUserServiceTest {
 
     @InjectMocks
     private AdminUserService adminUserService;
-
-    @Mock
-    private MatchRepository matchRepository;
 
     @Mock
     private UserRepository userRepository;
@@ -120,47 +114,6 @@ public class AdminUserServiceTest {
         // when & then
         assertThrows(GlobalException.class, () -> {
             adminUserService.getUserDetails(userId);
-        });
-    }
-
-    @Test
-    @DisplayName("사용자 포인트 지급 성공")
-    void adjustUserPoints_Success() {
-        // given
-        Long userId = 1L;
-        BigDecimal amountToAdd = new BigDecimal("5000");
-        PointAdjustRequest request = new PointAdjustRequest(amountToAdd, "베팅 승리 보상");
-
-        // User 엔티티 생성 시 Builder는 point를 1000으로 초기화(초기 지급 포인트)
-        User mockUser = User.builder()
-                .email("test@test.com")
-                .username("testuser")
-                .password("password")
-                .build();
-
-        given(userRepository.findById(userId)).willReturn(Optional.of(mockUser));
-
-        // when
-        PointAdjustResponse response = adminUserService.adjustUserPoints(userId, request);
-
-        // then
-        assertThat(response.userId()).isEqualTo(mockUser.getId());
-        assertThat(response.adjustedAmount()).isEqualTo(amountToAdd);
-        assertThat(response.finalBalance()).isEqualTo(new BigDecimal("6000"));
-        assertThat(mockUser.getPoint()).isEqualTo(new BigDecimal("6000"));
-    }
-
-    @Test
-    @DisplayName("사용자 포인트 지급 실패 - 존재하지 않는 사용자")
-    void adjustUserPoints_Fail_UserNotFound() {
-        // given
-        Long userId = 999L;
-        PointAdjustRequest request = new PointAdjustRequest(new BigDecimal("5000"), "이벤트 보상");
-        given(userRepository.findById(userId)).willReturn(Optional.empty());
-
-        // when & then
-        assertThrows(GlobalException.class, () -> {
-            adminUserService.adjustUserPoints(userId, request);
         });
     }
 }
